private routes and testing

+++ app.js
@@ -15,9 +15,10 @@ const PORT = process.env.PORT || 3000;
 app.use(bodyParser.json());
 
 
-app.post('/todos', (req,res) => {
+app.post('/todos',authenticate, (req,res) => {
     var todo = new Todo({
-        text: req.body.text
+        text: req.body.text,
+        owner_id: req.user._id
     });
 
     todo.save().then((todo) => {
@@ -27,20 +28,21 @@ app.post('/todos', (req,res) => {
     });
 });
 
-app.get('/todos', (req,res) => {
-    Todo.find({}).then((todos) => {
+app.get('/todos', authenticate, (req,res) => {
+    Todo.find({owner_id: req.user._id}).then((todos) => {
         res.send({todos})
     }).catch((err) => {
         res.status(400).send(err);
     });
 });
 
-app.get('/todos/:id', (req,res) => {
+app.get('/todos/:id', authenticate, (req,res) => {
     var id = req.params.id;
     if(!ObjectID.isValid(id)) {
         return res.status(404).send({error: "ID not valid"});
     }
-    Todo.findById(id).then((todo) => {
+
+    Todo.findOne({_id: id, owner_id: req.user._id}).then((todo) => {
         if(!todo) {
             return res.status(400).send({error: "User ID not found"});
         }
@@ -50,12 +52,12 @@ app.get('/todos/:id', (req,res) => {
     });
 });
 
-app.delete('/todos/:id', (req,res) => {
+app.delete('/todos/:id', authenticate,  (req,res) => {
     var id = req.params.id;
     if(!ObjectID.isValid(id)) {
         return res.status(404).send({error: "ID not valid"});
     }
-    Todo.findOneAndDelete({_id:id}).then((todo) => {
+    Todo.findOneAndDelete({_id:id, owner_id: req.user._id}).then((todo) => {
         if(!todo) {
             return res.status(400).send({error: "User ID not found"});
         }
@@ -65,7 +67,7 @@ app.delete('/todos/:id', (req,res) => {
     });
 });
 
-app.patch("/todos/:id", (req,res) => {
+app.patch("/todos/:id",authenticate, (req,res) => {
     var id = req.params.id;
     var body = _.pick(req.body,["text","completed"]);
     if(!ObjectID.isValid(id)) {
@@ -78,7 +80,7 @@ app.patch("/todos/:id", (req,res) => {
         body.completedAt = null;
     }
 
-    Todo.findOneAndUpdate({_id: id},{$set: body},{new:true}).then((todo) => {
+    Todo.findOneAndUpdate({_id: id, owner_id: req.user._id},{$set: body},{new:true}).then((todo) => {
         if(!todo) {
             return res.status(400).send({error: "User ID not found"});
         }
@@ -122,8 +124,16 @@ app.post("/users/login", (req,res) => {
             }
         });
     }).catch((err) => {
-        console.lo0g(err);
+        console.log(err);
         res.status(400).send(err);
+    });
+});
+
+app.delete("/users/me/token", authenticate, (req,res) => {
+    req.user.removeToken(req.token).then(() => {
+        res.status(200).send();
+    }).catch(() => {
+        res.status(400).send();
     });
 });
